Crate stream_cipher

This crate defines a set of traits which define functionality of stream ciphers.

See RustCrypto/stream-ciphers for ciphers implementation.

Re-exports

pub use generic_array;

pub use block_cipher;

Modules

consts

Type aliases for many constants.

Structs

InvalidKeyNonceLength

The error type returned when key and/or nonce used in stream cipher initialization had an invalid length.

LoopError

The error type returned when stream cipher has reached the end of a keystream.

OverflowError

The error type returned when a cipher position can not be represented by the requested type.

Traits

FromBlockCipher

Trait for initializing a stream cipher from a block cipher

NewStreamCipher

Stream cipher creation trait.

SeekNum

Trait implemented for numeric types which can be used with the SyncStreamCipherSeek trait.

StreamCipher

Stream cipher core trait which covers both synchronous and asynchronous ciphers.

SyncStreamCipher

Synchronous stream cipher core trait.

SyncStreamCipherSeek

Trait for seekable stream ciphers.

Type Definitions

Key

Key for an algorithm that implements NewStreamCipher.

Nonce

Nonce for an algorithm that implements NewStreamCipher.