Crate stream_cipher

This crate defines a set of traits which define functionality of stream ciphers.

See RustCrypto/stream-ciphers for ciphers implementation.

Re-exports

pub extern crate generic_array;

Structs

InvalidKeyNonceLength

Error which notifies that key or/and nonce used in stream cipher initialization had an invalid length.

LoopError

Error which notifies that stream cipher has reached the end of a keystream.

Traits

NewStreamCipher

Stream cipher creation trait.

StreamCipher

Stream cipher core trait which covers both synchronous and asynchronous ciphers.

SyncStreamCipher

Synchronous stream cipher core trait.

SyncStreamCipherSeek

Synchronous stream cipher seeking trait.